[FIRRTL] Fold MemToRegOfVec into FullReset
Inline comb-mem to reg-of-vec conversion into firrtl-full-reset and delete
the standalone firrtl-mem-to-reg-of-vec pass. Conversion remains limited to
modules in asynchronous full-reset domains.

Update unit tests to drive conversion via FullReset annotations and adjust
the FIRRTL rationale accordingly.
