+++ include/circt/Dialect/FIRRTL/Passes.td
@@ -542,26 +542,6 @@ def ExtractInstances : Pass<"firrtl-extract-instances", "firrtl::CircuitOp"> {
   ];
 }
 
-def MemToRegOfVec : Pass<"firrtl-mem-to-reg-of-vec", "firrtl::CircuitOp"> {
-  let summary = "Convert combinational memories to a vector of registers";
-  let description = [{
-    Utility pass that converts combinational memories in modules that appear
-    in the effective design into vectors of registers. Production pipelines
-    perform this conversion through `firrtl-full-reset` for modules in
-    asynchronous full-reset domains; this pass remains available for custom
-    pipelines and unit tests.
-  }];
-  let options = [
-    Option<"ignoreReadEnable", "ignore-read-enable-mem", "bool",
-                "false",
-    "ignore the read enable signal, instead of assigning X on read disable">
-   ];
-  let statistics = [
-    Statistic<"numConvertedMems", "num-converted-mems",
-      "Number of memories converted to registers">,
-  ];
-}
-
 def DropName : Pass<"firrtl-drop-names", "firrtl::FModuleOp"> {
   let summary = "Drop interesting names";
   let description = [{
