+++ lib/Dialect/FIRRTL/Transforms/InferResets.cpp
@@ -1,4 +1,4 @@
-//===- InferResets.cpp - Infer abstract reset types ------------*- C++ -*-===//
+//===----------------------------------------------------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -112,8 +112,6 @@ static T &operator<<(T &os, const ResetKind &kind) {
   }
   return os;
 }
-
-
 
 //===----------------------------------------------------------------------===//
 // Pass Infrastructure
@@ -957,4 +955,3 @@ LogicalResult InferResetsPass::verifyNoAbstractReset() {
     return failure();
   return success();
 }
-
